Add Device hook; Optimize Menu component; Fix style issue on Netlify
