+++ src/data/urls.ts
@@ -1,0 +1,32 @@
+const Urls = {
+  usTaxes: {
+    start: '/start'
+  },
+  taxPayer: {
+    root: '/taxpayer',
+    info: '/info',
+    spouseAndDependent: '/spouseanddependent',
+    contactInfo: '/contact'
+  },
+  refund: '/refundinfo',
+  questions: '/questions',
+  income: {
+    w2s: '/income/w2jobinfo',
+    f1099s: '/income/f1099s',
+    realEstate: '/income/realestate'
+  },
+  deductions: {
+    f1098es: '/deductions/studentloaninterest'
+  },
+  credits: {
+    main: '/credits',
+    eic: '/credits/eic'
+  },
+  createPdf: '/createpdf',
+  summary: '/summary',
+  default: ''
+}
+
+Urls.default = Urls.usTaxes.start
+
+export default Urls
